selenium Grid StandAlone

+++ src/test/java/testBase/BaseClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Date;
@@ -12,11 +13,14 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
@@ -25,6 +29,7 @@ public class BaseClass {
 	public static  WebDriver driver;
 	public Logger logger;
 	public Properties p;
+	DesiredCapabilities capabilities;
 	
 	@BeforeClass(groups= {"Sanity","Regression","Master"})
 	@Parameters({"os","browser"})
@@ -37,16 +42,49 @@ public class BaseClass {
 		
 		logger=LogManager.getLogger(this.getClass());
 		
-		switch(br.toLowerCase()) {
-		case "chrome" :driver = new ChromeDriver(); break;
-		case "edge" :driver = new EdgeDriver(); break;
-		case "firefox" :driver = new FirefoxDriver(); break;
-		default : System.out.println("Invalid browser name..");
 		
-		return ; 
-		
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote")){
+			capabilities = new DesiredCapabilities();
+			//os
+			if(os.equalsIgnoreCase("windows")) {
+				capabilities.setPlatform(Platform.WIN11);
+			}else if(os.equalsIgnoreCase("mac")) {
+				capabilities.setPlatform(Platform.MAC);
+			}else {
+				System.out.println("No matching os");
+				return;
+			}
+			
+			//browser
+			switch(br.toLowerCase()) {
+			case "crome":capabilities.setBrowserName("chrome"); break;
+			case "edge": capabilities.setBrowserName("MicrosoftEdge");
+			default: System.out.println("No matching browser"); return;
+			
+			}
+			
+			
+			
+			
+			
+			
+			
+			
 		}
 		
+		if(p.getProperty("execution_env").equalsIgnoreCase("local")) {
+			switch(br.toLowerCase()) {
+			case "chrome" :driver = new ChromeDriver(); break;
+			case "edge" :driver = new EdgeDriver(); break;
+			case "firefox" :driver = new FirefoxDriver(); break;
+			default : System.out.println("Invalid browser name..");
+			
+			return ; 
+			
+			}
+				driver = new RemoteWebDriver(new URL("http://localhost:4444/wb/hub"),capabilities);
+			
+		}
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
